Solved some bugs
